fix disappearing mouse on 64-bit architectures
Currently when building refind 64bit there is a bug that causes the
mouse to disappear shortly after moving it.

Fix it by changing TargetX and TargetY int variables back to 32bit for
both architectures.

The commit 4edad4 that changed them to INTN which resulted
in them being int64 on x86_64 introduced support for pointing devices on
32bit builds, and this commit doesn't revert that support, just
preserves earlier, good behaviour on 64bit.