Better icon for Haiku.
There were previous merge requests trying to replace the icon for Haiku,
but they were based on a 3D looking icon, which also isn't the official
Haiku logo.

This one uses Haiku official material (based on a sticker design) and
matches the look of other icons.

This was made using Haiku native image editor Wonderbrush because that's
the format where Haiku artwork is available. The SVG export is not
correct, in particular it doesn't render the shadow correctly. So I also
attached the Wonderbrush file. Sorry about using unusual formats, I hope
that isn't too much of a problem.

If you know how the shado was made in the other icons, maybe you can
take the SVG, remove the current shadow, and add one exactly matching
the other icons. I think the one I did in Wonderbrush is pretty close so
the PNG icon integrates well with the others.

I originally posted this to the forum, but was asked to do a merge
request instead. So here I am.